added staticData folder
