+++ src/components/navbar/Navbar.css
@@ -1,9 +1,3 @@
-<<<<<<< HEAD
-=======
-
-@import "../../App.css";
-
->>>>>>> 6553ff6047cf854b3337c9fca60748dcf8bcae9f
 .simple-navigation {
   background-color: var(--primary-color);
   padding: 1rem;
@@ -42,13 +36,6 @@
   color: var(--secondary-color);
 }
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-
-=======
->>>>>>> b3116ee69803f3ed10585e7a7f2317945025ec3e
-=======
->>>>>>> 6553ff6047cf854b3337c9fca60748dcf8bcae9f
 .icons > a{
   color: var(--secondary-color);
 }
@@ -91,13 +78,6 @@
   right: 0.7rem;
 }
 
-<<<<<<< HEAD
-<<<<<<< HEAD
-=======
-
->>>>>>> b3116ee69803f3ed10585e7a7f2317945025ec3e
-=======
->>>>>>> 6553ff6047cf854b3337c9fca60748dcf8bcae9f
 .sub-container {
   display: flex;
 }
