+++ src/pages/home/Home.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import "./Home.css";
-import { Navbar, Footer } from "../../components/componentIndex";
+import { Navbar, Footer } from "../../components";
 import {Main} from "./main/Main"
 
 export const Home = () =>{
