+++ src/pages/product/Product.css
@@ -1,35 +1,35 @@
-.container{
+.productContainer{
     display: grid;
     grid-template-columns: 1fr 5fr;
 }
 
-.drawer{
+.productContainer .drawer{
     width: 100%;
     padding: 1rem 0;
 }
 
-.drawer .heading{
+.productContainer .drawer .heading{
     margin: .5rem;
     text-align: left;
 }
 
-.main-content{
+.productContainer .main-content{
     display: flex;
     justify-content: space-evenly;
     flex-wrap: wrap;
 }
 
-.card{
+.productContainer .card{
     background-color: var(--white-section);
     padding: 0;
-    width: 12rem;
-    height: 15rem;
+    width: 15rem;
+    height: 16rem;
 }
-.card-text{
+.productContainer .card-text{
     padding: 0 1rem;
 }
 
-.img{
+.productContainer .img{
     object-fit: cover;
     height: 70%;
     border-radius: 1rem 1rem 0 0;
