+++ src/pages/wishlist/Wishlist.css
@@ -1,54 +1,56 @@
-.container{
+@import "../../App.css";
+
+.wishlistContainer{
     height: auto;
 }
 
-.sub-container{
+.wishlistContainer .sub-container{
     display: flex;
     justify-content: space-evenly;
 }
 
-.fa-shopping-cart{
+.wishlistContainer .fa-shopping-cart{
     padding: 0 0.5rem;
 }
 
-.favourite{
+.wishlistContainer .favourite{
     padding: 0 3rem;
 }
 
-.sub-container{
+.wishlistContainer .sub-container{
     align-self: flex-start;
-    width: 80%;
+    width: 100%;
     flex-wrap: wrap;
 }
 
-.card{
+.wishlistContainer .card{
     padding: 0;
     margin: 1rem;
     background-color: var(--white-section);
     height: 16rem;
 }
 
-.horizontal .img{
+.wishlistContainer .horizontal .img{
     width: 100%;
     height: 60%;
     object-fit: cover;
     border-radius: 1rem 1rem 0 0;
 }
 
-.card-footer .btn{
+.wishlistContainer .card-footer .btn{
     background-color: var(--primary-color);
     color: var(--secondary-color);
     padding: 0.2rem;
     border-radius: 0.5rem;
 }
 
-.icon{
+.wishlistContainer .icon{
     border: 1px solid var(--primary-color);
     padding: 0.2rem;
     border-radius: 0.5rem;
 }
 
-.icon .favourite{
+.wishlistContainer .icon .favourite{
     color: var(--primary-color);
     text-decoration: none;
 }
