+++ src/pages/wishlist/main/Main.jsx
@@ -2,7 +2,7 @@ import { Card } from "./Card";
 
 const Main = () => (
   <>
-    <section class="margin-0 container">
+    <section class="margin-0 wishlistContainer">
       <Card />
     </section>
   </>
